+++ src/main/java/insane96mcp/iguanatweaksexpanded/network/message/SyncITEEnchantingTableEnchantments.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksexpanded.network.message;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchanting.SREnchantingTableMenu;
+import insane96mcp.iguanatweaksexpanded.module.experience.enchanting.ITEEnchantingTableMenu;
 import insane96mcp.iguanatweaksexpanded.network.NetworkHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.FriendlyByteBuf;
@@ -15,14 +15,14 @@ import java.util.function.Supplier;
 
 import static insane96mcp.iguanatweaksexpanded.network.NetworkHandler.CHANNEL;
 
-public class SyncSREnchantingTableEnchantments {
+public class SyncITEEnchantingTableEnchantments {
 	List<EnchantmentInstance> enchantmentInstances;
 
-	public SyncSREnchantingTableEnchantments(List<EnchantmentInstance> enchantmentInstances) {
+	public SyncITEEnchantingTableEnchantments(List<EnchantmentInstance> enchantmentInstances) {
 		this.enchantmentInstances = enchantmentInstances;
 	}
 
-	public static void encode(SyncSREnchantingTableEnchantments pkt, FriendlyByteBuf buf) {
+	public static void encode(SyncITEEnchantingTableEnchantments pkt, FriendlyByteBuf buf) {
 		buf.writeShort(pkt.enchantmentInstances.size());
 		for (EnchantmentInstance enchantmentInstance : pkt.enchantmentInstances) {
 			buf.writeResourceLocation(ForgeRegistries.ENCHANTMENTS.getKey(enchantmentInstance.enchantment));
@@ -30,18 +30,18 @@ public class SyncSREnchantingTableEnchantments {
 		}
 	}
 
-	public static SyncSREnchantingTableEnchantments decode(FriendlyByteBuf buf) {
+	public static SyncITEEnchantingTableEnchantments decode(FriendlyByteBuf buf) {
 		List<EnchantmentInstance> enchantments = new ArrayList<>();
 		int size = buf.readShort();
 		for (int i = 0; i < size; i++) {
 			enchantments.add(new EnchantmentInstance(ForgeRegistries.ENCHANTMENTS.getValue(buf.readResourceLocation()), buf.readShort()));
 		}
-		return new SyncSREnchantingTableEnchantments(enchantments);
+		return new SyncITEEnchantingTableEnchantments(enchantments);
 	}
 
-	public static void handle(final SyncSREnchantingTableEnchantments message, Supplier<NetworkEvent.Context> ctx) {
+	public static void handle(final SyncITEEnchantingTableEnchantments message, Supplier<NetworkEvent.Context> ctx) {
 		ctx.get().enqueueWork(() -> {
-			if (NetworkHelper.getSidedPlayer(ctx.get()).containerMenu instanceof SREnchantingTableMenu enchantingTableMenu) {
+			if (NetworkHelper.getSidedPlayer(ctx.get()).containerMenu instanceof ITEEnchantingTableMenu enchantingTableMenu) {
 				enchantingTableMenu.updateEnchantmentsChosen(message.enchantmentInstances);
 			}
 		});
@@ -49,7 +49,7 @@ public class SyncSREnchantingTableEnchantments {
 	}
 
 	public static void sync(List<EnchantmentInstance> enchantmentInstances) {
-		Object msg = new SyncSREnchantingTableEnchantments(enchantmentInstances);
+		Object msg = new SyncITEEnchantingTableEnchantments(enchantmentInstances);
 		CHANNEL.sendTo(msg, Minecraft.getInstance().player.connection.getConnection(), NetworkDirection.PLAY_TO_SERVER);
 	}
 }